move param checks in a separate file
